add tpu-mlir forward inference for qat train

+++ sophgo_mq/convert_deploy.py
@@ -1,6 +1,9 @@
+import io
 import json
 import os.path as osp
 import os
+import pdb
+import subprocess
 import numpy as np
 import torch
 from torch.fx import GraphModule
@@ -43,6 +46,7 @@ from sophgo_mq.fake_quantize import (
 )
 # from utils.mlir_shell import mlir_opt_for_top, mlir_lowering, mlir_to_model, f32_blobs_compare
 # from tools.model_runner import mlir_inference, free_mlir_module
+from tools.model_transform import model_transform_func
 import torchvision.transforms as TorchTransforms
 
 FP8_FAKEQUANTIZER = [E4M3FakeQuantize, E5M2FakeQuantize]
@@ -52,11 +56,12 @@ INT_FAKEQUANTIZER = [LearnableFakeQuantize, NNIEFakeQuantize, FixedFakeQuantize,
 
 __all__ = ['convert_deploy']
 
+model_onnx_mem = None
 @register_deploy_function("CNN")
 def convert_merge_bn(model: GraphModule, **kwargs):
     # print('wlog before convert_merge_bn, model.named_modules:', dict(model.named_modules())[''])
     # print('wlog before convert_merge_bn, model.graph:', model.graph)
-    logger.info("Merge BN for deploy.")
+    # logger.info("Merge BN for deploy.")
     nodes = list(model.graph.nodes)
     modules = dict(model.named_modules())
     for node in nodes:
@@ -69,10 +74,12 @@ def convert_merge_bn(model: GraphModule, **kwargs):
 
 @register_deploy_function("CNN")
 def convert_onnx(model: GraphModule, input_shape_dict, dummy_input, onnx_model_path, **kwargs):
-    pt_file_name = onnx_model_path.split('.')
-    pt_file_name[-1] = 'pt'
+    # pt_file_name = onnx_model_path.split('.')
+    # pt_file_name[-1] = 'pt'
     #torch.save(model, '.'.join(pt_file_name))
-    logger.info("Export to onnx, onnx_model_path:{}".format(onnx_model_path))
+    export_to_mem = 'not_gen_bmodel' in kwargs and kwargs['not_gen_bmodel']
+    if not export_to_mem:
+        logger.info("Export to onnx, onnx_model_path:{}".format(onnx_model_path))
     model = model.cpu()
     output_names = kwargs.get('output_names', [])
     dynamic_axes = kwargs.get('dynamic_axes', {})
@@ -88,7 +95,8 @@ def convert_onnx(model: GraphModule, input_shape_dict, dummy_input, onnx_model_p
     
     # open all fake quant node to export
     if isinstance(model, torch.fx.graph_module.GraphModule):
-        print(">>>>> print graphmodule before export", model)
+        if not export_to_mem:
+            print(">>>>> print graphmodule before export", model)
         # print(">>>>> print graph before export")
         # model.graph.print_tabular()
         for name, submodule in model.named_modules():
@@ -100,7 +108,10 @@ def convert_onnx(model: GraphModule, input_shape_dict, dummy_input, onnx_model_p
             class_of_submodule = submodule.__class__
             if class_of_submodule in FP8_FAKEQUANTIZER:
                 quant_mode = "FP8"
+    if export_to_mem:
+        onnx_model_path = io.BytesIO()
 
+    simplified_model = None
     with torch.no_grad():
         try:
             from torch.onnx.utils import ONNXCheckerError
@@ -122,15 +133,26 @@ def convert_onnx(model: GraphModule, input_shape_dict, dummy_input, onnx_model_p
                                 opset_version=opset_version,
                                 do_constant_folding=True,
                                 custom_opsets={'Sophgo_custom' : opset_version})
-            tmp_model = onnx.load(onnx_model_path)
+            if export_to_mem:
+                tmp_model = onnx.load_model_from_string(onnx_model_path.getvalue())
+            else:
+                tmp_model = onnx.load(onnx_model_path)
             simplified_model, check = simplify(tmp_model)
-            onnx.save_model(simplified_model, onnx_model_path)
+            if not export_to_mem:
+                onnx.save_model(simplified_model, onnx_model_path)
 
-    model_onnx = onnx.load(onnx_model_path)
+    if export_to_mem:
+        model_onnx = simplified_model if simplified_model is not None else onnx_model_path
+    else:
+        model_onnx = onnx.load(onnx_model_path)
     onnx.checker.check_model(model_onnx)
     model_onnx = onnx.shape_inference.infer_shapes(model_onnx)
-    os.system(f"rm -f {onnx_model_path}")
-    onnx.save(model_onnx, onnx_model_path)
+    if export_to_mem:
+        global model_onnx_mem
+        model_onnx_mem = model_onnx
+    else:
+        os.system(f"rm -f {onnx_model_path}")
+        onnx.save(model_onnx, onnx_model_path)
 
      
 @register_deploy_function("Transformer")
@@ -360,8 +382,9 @@ def deploy_qparams_Academic_NLP(model: GraphModule, onnx_model_path, model_name,
 def deploy_qparams_sophgo_tpu(model: GraphModule, onnx_model_path, model_name, quant_type_dict, **kwargs):
     logger.info("Extract qparams for sophgo_tpu.")
     cali_mode = "sophgo_tpu"
-    remove_fakequantize_and_collect_params_sophgo(onnx_model_path, model_name, quant_type_dict)
-    print("导出calitable")
+    global model_onnx_mem
+    model_onnx_mem = remove_fakequantize_and_collect_params_sophgo(onnx_model_path, model_name, quant_type_dict, model_onnx_mem)
+    # print("导出calitable")
     output_path = osp.dirname(onnx_model_path)
     context_filename = osp.join(output_path, '{}_clip_ranges.json'.format(model_name))
     file_h = open(context_filename, "r")
@@ -472,6 +495,7 @@ def convert_deploy(model: GraphModule, net_type='CNN',
         if val_loader == None:
             print(f'VAL Loader not set for deploy!')
             return
+        export_to_mm = 'not_gen_bmodel' in kwargs and kwargs['not_gen_bmodel']
         mlir_scale = '1,1,1'
         mlir_mean = '0,0,0'
         transforms = val_loader.dataset.transform.transforms
@@ -484,28 +508,46 @@ def convert_deploy(model: GraphModule, net_type='CNN',
                 mlir_scale = ','.join([str(round(i,4)) for i in mlir_scale.tolist()])
                 mlir_mean = ','.join([str(round(i,4)) for i in mlir_mean.tolist()])
 
-        onnx_filename = os.path.join(output_path, '{}_deploy_model.onnx'.format(model_name))
+        if not export_to_mm:
+            onnx_filename = os.path.join(output_path, '{}_deploy_model.onnx'.format(model_name))
+        else:
+            global model_onnx_mem
+            onnx_filename = model_onnx_mem
         shape_str_list = []
         for name in input_shape_dict:
             shape_str = ','.join([str(i) for i in input_shape_dict[name]])
             shape_str_list.append(f'[{shape_str}]')
         shape_str_list = ','.join(shape_str_list)
-        cmd_str = f"model_transform.py \
-        --model_name {model_name}_qat \
-        --model_def {onnx_filename} \
-        --input_shapes [{shape_str_list}] \
-        --mean {mlir_scale} \
-        --scale {mlir_mean} \
-        --keep_aspect_ratio \
-        --pixel_format rgb \
-        --mlir {model_name}_qat.mlir"
-        print('cmd_str:', cmd_str)
-        os.system(cmd_str)
+        
+        if export_to_mm:
+            model_transform_func(f'{model_name}_qat', 
+                                onnx_filename, list(input_shape_dict.values()), 
+                                mlir_scale, mlir_mean,
+                                f'{model_name}_qat.mlir')
+        else:
+            cmd_str = f"model_transform.py \
+            --model_name {model_name}_qat \
+            --model_def {onnx_filename} \
+            --input_shapes [{shape_str_list}] \
+            --mean {mlir_scale} \
+            --scale {mlir_mean} \
+            --keep_aspect_ratio \
+            --pixel_format rgb \
+            --mlir {model_name}_qat.mlir"
+            os.system(cmd_str)
 
         calibration_table = os.path.join(output_path, '{}_cali_table_from_sophgo_mq_sophgo_tpu'.format(model_name))
         test_input = os.path.join(output_path, 'input_data_0.npz')
         test_reference = os.path.join(output_path, 'layer_outputs_0.npz')
         quantize_table = ''
+        not_gen_bmodel = ''
+        test_input, test_reference = '', ''
+        if export_to_mm:
+            not_gen_bmodel = '--not_gen_bmodel'
+        else:
+            test_input = f'--test_input {test_input}'
+            test_reference = f'--test_reference {test_reference}'
+
         quantize_mode = 'INT8'
         quantize_str = 'int8_sym'
         if 'bf16_mix_prec' in kwargs and kwargs['bf16_mix_prec']:
@@ -518,11 +560,9 @@ def convert_deploy(model: GraphModule, net_type='CNN',
         --mlir {model_name}_qat.mlir \
         --quantize {quantize_mode} \
         --calibration_table {calibration_table} {quantize_table} \
-        --chip {chip} \
-        --test_input {test_input} \
-        --test_reference {test_reference} \
+        --chip {chip} {test_input} {test_reference}\
         --fazzy_match \
-        --tolerance 0.99,0.90 --debug \
+        --tolerance 0.99,0.90 --debug {not_gen_bmodel}\
         --model {model_name}_qat_{chip.lower()}_{quantize_str}.{bmodel_ext}")
         return f'./{model_name}_qat_{chip.lower()}_{quantize_str}_tpu.mlir'
